pam file, ConstructorTest, RegistrationTest changed

+++ src/test/java/uitests/RegistrationTest.java
@@ -5,6 +5,7 @@ import api.user.User;
 import api.user.UserGenerator;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import io.qameta.allure.Description;
+import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import org.apache.commons.lang3.RandomStringUtils;
@@ -21,13 +22,16 @@ import static org.junit.Assert.assertTrue;
 
 public class RegistrationTest {
     private WebDriver driver;
-
     private User user;
+    private LoginPage loginPage;
 
     @Before
     public void startUp() {
         WebDriverManager.chromedriver().setup();
         user = UserGenerator.getRandom();
+        System.setProperty("webdriver.chrome.driver", "src\\test\\resources\\yandexdriver.exe");
+        driver = new ChromeDriver();
+        driver.get("https://stellarburgers.nomoreparties.site/");
     }
 
     @After
@@ -36,27 +40,37 @@ public class RegistrationTest {
         UserClient userClient = new UserClient();
         ValidatableResponse response = userClient.login(user);
         String token = response.extract().path("accessToken");
-        userClient.delete(token);
+        if (token != null) userClient.delete(token);
+    }
+
+    @Step("Переход на страницу регистрации")
+    public void goToRegistrationTab() {
+        MainPage objMainPage = new MainPage(driver);
+        objMainPage.pressOnPersonalAccount();
+
+        loginPage = new LoginPage(driver);
+        loginPage.pressOnRegisterButton();
+
+    }
+
+    @Test
+    @DisplayName("Регистрация пользователя, негативный кейс")
+    @Description("Проверка ошибки для некорректного пароля")
+    public void registrationNegativeTest() {
+        goToRegistrationTab();
+        RegisterPage registerPage = new RegisterPage(driver);
+        registerPage.inputAllFieldsAndPressRegistrationButton(user.getName(), user.getEmail(), RandomStringUtils.randomAlphabetic(3));
+        assertTrue("Ожидается сообщение о неитьоподходящем пароле", registerPage.checkMessageAboutNotSuitablePasswordAppear());
+
     }
 
     @Test
     @DisplayName("Регистрация пользователя")
-    @Description("Проверка ошибки для некорректного пароля и умпешной регистрации")
+    @Description("Проверка уcпешной регистрации")
     public void registrationTest() {
-        System.setProperty("webdriver.chrome.driver", "src\\test\\resources\\yandexdriver.exe");
-        driver = new ChromeDriver();
-        driver.get("https://stellarburgers.nomoreparties.site/");
-        MainPage objMainPage = new MainPage(driver);
-        objMainPage.pressOnPersonalAccount();
-
-        LoginPage loginPage = new LoginPage(driver);
-        loginPage.pressOnRegisterButton();
-
+        goToRegistrationTab();
         RegisterPage registerPage = new RegisterPage(driver);
-
-        registerPage.inputAllFieldsAndPressRegistrationButton(user.getName(), user.getEmail(), RandomStringUtils.randomAlphabetic(3));
-        assertTrue(registerPage.checkMessageAboutNotSuitablePasswordAppear());
-        registerPage.inputPassword(user.getPassword());
+        registerPage.inputAllFieldsAndPressRegistrationButton(user.getName(), user.getEmail(), user.getPassword());
         registerPage.pressRegistrationButton();
         loginPage.waitLoginButtonToAppear();
         assertTrue("После успешной регистрации пользователь должен перенаправиться на окно со входом", loginPage.checkLoginButtonAppeared());
